Move IOHandler to separate file
